niffler-e-2-e-tests-python/requirements.txt
selene==2.0.0rc9
pytest
requests
python-dotenv
faker
urllib3
python-dotenv
sqlalchemy
sqlmodel
psycopg2-binary
curlify
confluent-kafka
allure-pytest
pkce
jinja2